Allow "Your profile" link to be found by different link text
We're changing the "Your profile" link to "Your account". This change is
to temporarily check for either link text on the page so that the
functional tests don't fail after the change has been deployed. Once the
link text change is live, we can simplify this again to only look for
the new wording.