Only check for new link to the profile page
In #628020a7845616c1ffa01ff7e9dae0f21de7edcf we started to check for the
profile link by looking for the text "Your profile" or "Your account".
This was to avoid the tests breaking while we changed the link text. Now
that the link text has been updated in production we can simplify the code
by only checking for the new version of the text.