upgrade setuptools
the official base python images now ship with incompatible versions of
wheel and setuptools, which can cause issues with running
`python setup.py bdist_wheel`, which happens for some of our
dependencies and subdependencies.

For now, a workaround is to bump setuptools to a version that doesn't
have this incompatibility. We should remove this line in future once
the base python image has been updated