increase vertical size of functional tests chromedriver
we've been seeing some issues with the sticky nav on template page,
where we choose a template type to create but when we try to hit the
continue button we scrollIntoViewIfNeeded and then when we click the
button it misses. set quite a tall window in the hope that we don't need
to scroll so the tests are less flaky

if we're still seeing issues after this, we may need to introduce some
kind of sleep between the scrollIntoView and the click to let the sticky
nav transition to being part of the scrollable DOM, possibly adjust
the offset of the scroll to ensure the button is fully visible, or use
js to directly trigger the click event on the button.