template-folder manipulation tests: put in same xdist_group
due to what is thought to be a bug in the way we handle cache
consistency, these tests interfere with each other, so for
now ensure they won't run together. once that bug is fixed,
we should remove this restriction.

this is probably not the only cache-consistency-related failure
we'll see, but it is a notable one.