Dockerfile: Revert to Debian bookworm
At some point since this image was last built, the default Debian
flavour of the python:3.11 image has changed to Trixie. Unfortunately
we're pinning an old version of Chromium because newer versions cause
our functional tests to hang. Until we can fix that incompatibility,
we'll need to keep using Bookworm.